refactored right panel ui
